Merged revision 5 from XE branch
Log message of revision 5:
Fixed problem with size of  "Recent comments" button. Text is truncated on French and German (maybe on JA too).
Add project to a local repo AVs in module 'libsvn_repos-1.dll', also fixed crash when leaving a trailing slash.
If you make an error specifying the repository and try to commit, the "Import" pane disappears.